Total redesign of the RMT driver using a separate class that handles all of the interactions with the RMT peripheral. This design gets around the problem that methods in template classes cannot have the IRAM_ATTR attribute, which is crucial for the interrupt handler.
